Mirror module layout in folder layout.
