+++ lib/json_p3.rb
@@ -1,10 +1,11 @@
 # frozen_string_literal: true
 
 require_relative "json_p3/version"
-require_relative "json_p3/environment"
+require_relative "json_p3/errors"
+require_relative "json_p3/cache"
+require_relative "json_p3/path/environment"
 require_relative "json_p3/pointer"
 require_relative "json_p3/patch"
-require_relative "json_p3/lexer"
 
 # RFC 9535 JSONPath query expressions for JSON.
 module JSONP3
