+++ lib/json_p3/path/filter.rb
