+++ lib/json_p3/path/function.rb
