+++ lib/json_p3/path/function_extensions/pattern.rb
