+++ lib/json_p3/path/function_extensions/search.rb
@@ -1,13 +1,11 @@
 # frozen_string_literal: true
 
-require_relative "../cache"
-require_relative "../function"
 require_relative "pattern"
 
 module JSONP3
   module Path
-    # The standard `match` function.
-    class Match < FunctionExtension
+    # The standard `search` function.
+    class Search < FunctionExtension
       ARG_TYPES = %i[value_expression value_expression].freeze
       RETURN_TYPE = :logical_expression
 
@@ -29,9 +27,9 @@ module JSONP3
         return false unless pattern.is_a?(String) && value.is_a?(String)
 
         if @cache_size.positive?
-          re = @cache[pattern] || Regexp.new(full_match(pattern))
+          re = @cache[pattern] || Regexp.new(JSONP3::Path.map_iregexp(pattern))
         else
-          re = Regexp.new(full_match(pattern))
+          re = Regexp.new(JSONP3::Path.map_iregexp(pattern))
           @cache[pattern] = re
         end
 
@@ -40,24 +38,6 @@ module JSONP3
         raise if @raise_errors
 
         false
-      end
-
-      private
-
-      def full_match(pattern)
-        parts = [] # : Array[String]
-        explicit_caret = pattern.start_with?("^")
-        explicit_dollar = pattern.end_with?("$")
-
-        # Replace '^' with '\A' and '$' with '\z'
-        pattern = pattern.sub("^", "\\A") if explicit_caret
-        pattern = "#{pattern[..-1]}\\z" if explicit_dollar
-
-        # Wrap with '\A' and '\z' if they are not already part of the pattern.
-        parts << "\\A(?:" if !explicit_caret && !explicit_dollar
-        parts << JSONP3::Path.map_iregexp(pattern)
-        parts << ")\\z" if !explicit_caret && !explicit_dollar
-        parts.join
       end
     end
   end
