+++ lib/json_p3/path/node.rb
