+++ lib/json_p3/path/query.rb
