+++ lib/json_p3/path/segment.rb
