+++ lib/json_p3/path/selector.rb
