+++ lib/json_p3/path/serialize.rb
