+++ sig/json_p3/path/function.rbs
