+++ sig/json_p3/path/node.rbs
