+++ sig/json_p3/path/query.rbs
