+++ sig/json_p3/path/segment.rbs
