+++ sig/json_p3/path/serialize.rbs
